Fix rclone move command
removed for loop and running a single move command on the SavedClips directory. Rclone docs explain that a move command moves the contents of the source directory, NOT the source directory itself.
Also getting actual file count for push notification, instead of directory count. This only works with Tesla firmware since 2019.5...