Manually 'trim' freespace in the camera and music images.
The camera and music images end up with several thousand XFS extents. The
image snapshot creation time takes longer with more extents which causes
the snapshot operation to cause the car to see the exposed drive as "too slow"
when a snapshot happens. This is related to marcone/teslausb#239 and appears
to help.