Only add boot options to cmdline.txt once. (#350)
* Only add boot options to cmdline.txt once.

Running the setup script re-runs the make-root-fs-readonly.sh script
which appends the options 'fastboot', 'noswap', and 'ro' to the cmdline.txt
file. Adding those options repeatedly eventually causes the command line
to get long enough that the Pi won't boot.

This change makes a check for the option in the cmdline.txt before adding
the option so that the command line does not grow without bounds.