Rework LED flashing during setup
Instead of using /etc/stage_flash, use a function in setup-teslausb,
and make it correctly handle the difference in polarity between Pi0
and Pi4.

Only flash when setup-teslausb is not run interactively from a shell.

Since a lot of setup still happened after the documented "setup
complete" 5-flash stage, move the 5-flash notification to the end,
when setup is actually done. This has the added advantage that rc.local
may be able to continue a setup that was interrupted during the final
stage.

Remove the setup_led_on/off functions, since they weren't useful
considering that stage_flash already turned the led off after flashing.