enabled=true for notification services
When people want to enable a notification service, it makes more sense to me that the service is enabled after uncommenting the setting, by the default *_enabled=true, instead of the current *_enabled=false. 
This change also makes it consistent, as the AWS SNS notification setting already has the enabled=true value.